requirements.txt
numpy
matplotlib
h5py
PyYAML
SciencePlots
anesthetic
mpi4py
